fix: bind price_cents as i32 for referral_commission INSERT (sqlx Postgres)

+++ src/handlers/campaigns.rs
@@ -649,7 +649,7 @@ async fn campaign_settle(
              VALUES ($1, $2, 0.40) RETURNING id",
         )
         .bind(referrer_uid)
-        .bind(price_cents)
+        .bind(price_cents as i32)
         .fetch_one(&state.db_pool)
         .await;
